requirements.txt
asgiref==3.7.2
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
numpy==1.25.2
opencv-python==4.8.0.76
packaging==23.1
Pillow==10.0.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
